Created scripts.js
scripts.js handles all of the populate requests that are made by the user by pressing any category button. it then uses the respective php files in order to populate the articles from the database

You will need to use XAMPP or some other PHP enabled webserver provider/emulator in order to be able to see this work on your system.

Various files made in order to make this as effective and efficient as possible. This was also done with organization in mind.

Please contact Noah with any questions.