Update and rename scripts.js to script.js
Pretty substantial update.

Removed need to use foreach as it did not work with the outdated CSUSM webserver, and instead uses a for loop when retrieving the article JSON data from referenced PHP files.

Large update to ensure compatibility with the CSUSM web server. Works as intended. Future update to include compatibility for images and publish date, potentially other features?